modifying signature of big integer

+++ database/migrations/2022_04_30_091815_create_categories_movies_table.php
@@ -15,8 +15,8 @@ class CreateCategoriesMoviesTable extends Migration
     {
         Schema::create('categories_movies', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('movies_id');
-            $table->unsignedBigInteger('categories_id');
+            $table->unsignedInteger('movies_id');
+            $table->unsignedInteger('categories_id');
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent();
